requirements.txt
Flask
pyotp
pandas
supabase
python-dotenv
gunicorn